feat(as-5344): as-5344 refactoring householder appeal validation schema
