+++ packages/business-rules/src/schemas/components/boolean-validation.js
@@ -1,4 +1,4 @@
-const pinsYup = require('../../../lib/pins-yup');
+const pinsYup = require('../../lib/pins-yup');
 
 const booleanValidation = () => {
 	return pinsYup.bool().nullable();
