+++ packages/business-rules/src/schemas/components/section-state-validation.js
@@ -1,5 +1,5 @@
-const pinsYup = require('../../../lib/pins-yup');
-const { SECTION_STATE } = require('../../../constants');
+const pinsYup = require('../../lib/pins-yup');
+const { SECTION_STATE } = require('../../constants');
 
 const sectionStateValidation = () => {
 	return pinsYup.string().oneOf(Object.values(SECTION_STATE)).default('NOT STARTED');
