--- packages/business-rules/src/schemas/householder-appeal/appeal-schema/created-at/created-at-validation.js
@@ -1,7 +1,0 @@
-const dateValidation = require('../../generic-validators/date-validation');
-
-const createdAtValidation = () => {
-	return dateValidation(true);
-};
-
-module.exports = createdAtValidation;
