--- packages/business-rules/src/schemas/householder-appeal/appeal-schema/decision-date/decision-date-validation.js
@@ -1,7 +1,0 @@
-const dateValidation = require('../../generic-validators/date-validation');
-
-const decisionDateValidation = () => {
-	return dateValidation();
-};
-
-module.exports = decisionDateValidation;
