+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/eligibility/has-prior-approval-for-existing-home/prior-approval-validation.js
@@ -1,4 +1,4 @@
-const booleanValidation = require('../../../generic-validators/boolean-validation');
+const booleanValidation = require('../../../../components/boolean-validation');
 
 const priorApprovalValidation = () => {
 	return booleanValidation();
