+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/eligibility/householder-planning-persmission/householder-planning-permission-validation.js
@@ -1,4 +1,4 @@
-const booleanValidation = require('../../../generic-validators/boolean-validation');
+const booleanValidation = require('../../../../components/boolean-validation');
 
 const householderPlanningPermissionValidation = () => {
 	return booleanValidation();
