+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/index.js
@@ -9,38 +9,38 @@ const sectionStatesValidation = require('./section-states/section-states-validat
 const typeOfPlanningApplicationValidation = require('./type-of-planning-application/type-of-planning-application-validation');
 const appealTypeValidation = require('./appeal-type/appeal-type-validation');
 const stateValidation = require('./state/state-validation');
-const submissionDateValidation = require('./submission-date/submission-date-validation');
-const createdAtValidation = require('./created-at/created-at-validation');
-const updatedAtValidation = require('./updated-at/updated-at-validation');
-const decisionDateValidation = require('./decision-date/decision-date-validation');
 const planningApplicationNumberValidation = require('./planning-application-number/planning-application-number-validation');
 const lpaCodeValidation = require('./lpa-code/lpa-code-validation');
 const horizonIdValidation = require('./horizon-id/horizon-id-validation');
 const idValidation = require('./id/id-validation');
 const eligibilityValidation = require('./eligibility/eligibility-validation');
+const dateValidation = require('../../components/date-validation');
 
 const appealValidationSchema = () => {
-	return pinsYup.object().noUnknown(true).shape({
-		id: idValidation(),
-		horizonId: horizonIdValidation(),
-		lpaCode: lpaCodeValidation(),
-		planningApplicationNumber: planningApplicationNumberValidation(),
-		decisionDate: decisionDateValidation(),
-		createdAt: createdAtValidation(),
-		updatedAt: updatedAtValidation(),
-		submissionDate: submissionDateValidation(),
-		state: stateValidation(),
-		appealType: appealTypeValidation(),
-		typeOfPlanningApplication: typeOfPlanningApplicationValidation(),
-		email: emailValidation(),
-		eligibility: eligibilityValidation(),
-		aboutYouSection: aboutYouValidation(),
-		requiredDocumentsSection: requiredDocumentsValidation(),
-		yourAppealSection: yourAppealValidation(),
-		appealSubmission: appealSubmissionValidation(),
-		appealSiteSection: appealSiteValidation(),
-		sectionStates: sectionStatesValidation()
-	});
+	return pinsYup
+		.object()
+		.noUnknown(true)
+		.shape({
+			id: idValidation(),
+			horizonId: horizonIdValidation(),
+			lpaCode: lpaCodeValidation(),
+			planningApplicationNumber: planningApplicationNumberValidation(),
+			decisionDate: dateValidation(),
+			createdAt: dateValidation(true),
+			updatedAt: dateValidation(true),
+			submissionDate: dateValidation(),
+			state: stateValidation(),
+			appealType: appealTypeValidation(),
+			typeOfPlanningApplication: typeOfPlanningApplicationValidation(),
+			email: emailValidation(),
+			eligibility: eligibilityValidation(),
+			aboutYouSection: aboutYouValidation(),
+			requiredDocumentsSection: requiredDocumentsValidation(),
+			yourAppealSection: yourAppealValidation(),
+			appealSubmission: appealSubmissionValidation(),
+			appealSiteSection: appealSiteValidation(),
+			sectionStates: sectionStatesValidation()
+		});
 };
 
 module.exports = appealValidationSchema;
