+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/sections/about-you/your-details/your-details-validation.js
@@ -1,5 +1,5 @@
 const pinsYup = require('../../../../../../lib/pins-yup');
-const booleanValidation = require('../../../../generic-validators/boolean-validation');
+const booleanValidation = require('../../../../../components/boolean-validation');
 const appealingOnBehalfOfValidation = require('./appealing-on-behalf-of/appealing-on-behalf-of-validation');
 const nameValidation = require('./name/name-validation');
