+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/sections/appeal-site/health-and-safety/health-and-safety-validation.js
@@ -1,5 +1,5 @@
 const pinsYup = require('../../../../../../lib/pins-yup');
-const booleanValidation = require('../../../../generic-validators/boolean-validation');
+const booleanValidation = require('../../../../../components/boolean-validation');
 
 const healthAndSafetyValidation = () => {
 	return pinsYup
