+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/sections/appeal-submission/appeal-pdf-statement/appeal-pdf-statement-validation.js
@@ -1,0 +1,13 @@
+const pinsYup = require('../../../../../../lib/pins-yup');
+const uploadedFileValidation = require('../../../../../components/uploadedFileValidation');
+
+const appealPdfStatementValidation = () => {
+	return pinsYup
+		.object()
+		.shape({
+			uploadedFile: uploadedFileValidation()
+		})
+		.noUnknown(true);
+};
+
+module.exports = appealPdfStatementValidation;
