+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/sections/appeal-submission/appeal-submission-validation.js
@@ -1,23 +1,9 @@
 const pinsYup = require('../../../../../lib/pins-yup');
+const appealPdfStatementValidation = require('./appeal-pdf-statement/appeal-pdf-statement-validation');
 
 const appealSubmissionValidation = () => {
 	return pinsYup.object().shape({
-		appealPDFStatement: pinsYup
-			.object()
-			.shape({
-				uploadedFile: pinsYup
-					.object()
-					.shape({
-						id: pinsYup.string().trim().uuid().nullable().default(null),
-						name: pinsYup.string().trim().max(255).ensure(),
-						fileName: pinsYup.string().trim().max(255).ensure(),
-						originalFileName: pinsYup.string().trim().max(255).ensure(),
-						location: pinsYup.string().trim().nullable(),
-						size: pinsYup.number().nullable()
-					})
-					.noUnknown(true)
-			})
-			.noUnknown(true)
+		appealPDFStatement: appealPdfStatementValidation()
 	});
 };
