+++ packages/business-rules/src/schemas/householder-appeal/appeal-schema/sections/required-documents/required-documents-validation.js
@@ -1,4 +1,5 @@
 const pinsYup = require('../../../../../lib/pins-yup');
+const uploadedFileValidation = require('../../../../components/uploadedFileValidation');
 
 const requiredDocumentsValidation = () => {
 	return pinsYup
@@ -7,33 +8,13 @@ const requiredDocumentsValidation = () => {
 			originalApplication: pinsYup
 				.object()
 				.shape({
-					uploadedFile: pinsYup
-						.object()
-						.shape({
-							id: pinsYup.string().trim().uuid().nullable().default(null),
-							name: pinsYup.string().trim().max(255).ensure(),
-							fileName: pinsYup.string().trim().max(255).ensure(),
-							originalFileName: pinsYup.string().trim().max(255).ensure(),
-							location: pinsYup.string().trim().nullable(),
-							size: pinsYup.number().nullable()
-						})
-						.noUnknown(true)
+					uploadedFile: uploadedFileValidation()
 				})
 				.noUnknown(true),
 			decisionLetter: pinsYup
 				.object()
 				.shape({
-					uploadedFile: pinsYup
-						.object()
-						.shape({
-							id: pinsYup.string().trim().uuid().nullable().default(null),
-							name: pinsYup.string().trim().max(255).ensure(),
-							fileName: pinsYup.string().trim().max(255).ensure(),
-							originalFileName: pinsYup.string().trim().max(255).ensure(),
-							location: pinsYup.string().trim().nullable(),
-							size: pinsYup.number().nullable()
-						})
-						.noUnknown(true)
+					uploadedFile: uploadedFileValidation()
 				})
 				.noUnknown(true)
 		})
