--- packages/business-rules/src/schemas/householder-appeal/appeal-schema/submission-date/submission-date-validation.js
@@ -1,7 +1,0 @@
-const dateValidation = require('../../generic-validators/date-validation');
-
-const submissionDateValidation = () => {
-	return dateValidation();
-};
-
-module.exports = submissionDateValidation;
