--- packages/business-rules/src/schemas/householder-appeal/appeal-schema/updated-at/updated-at-validation.js
@@ -1,7 +1,0 @@
-const dateValidation = require('../../generic-validators/date-validation');
-
-const updatedAtValidation = () => {
-	return dateValidation(true);
-};
-
-module.exports = updatedAtValidation;
